requirements.txt
accelerate 
einops 
ffmpeg-python 
jupyter
ipykernel 
ruff 
scipy 
timm
opencv-python
datasets
matplotlib
torchmetrics
torchinfo
torchdata
torchmetrics
diffusers
transformers
polars
lightning
tensorboard
tensorboardX
segmentation-models-pytorch
fightingcv-attention